DRY is_trajectory_file and move (io/parse.ts -> trajectory/parse.ts)
