+++ tests/vitest/trajectory/parse.test.ts
@@ -1,12 +1,13 @@
 import {
   get_unsupported_format_message,
+  is_trajectory_file,
   parse_trajectory_data,
 } from '$lib/trajectory/parse'
 import { existsSync, readdirSync, readFileSync, statSync } from 'node:fs'
 import { join } from 'node:path'
 import process from 'node:process'
 import { gunzipSync } from 'node:zlib'
-import { describe, expect, it } from 'vitest'
+import { describe, expect, it, test } from 'vitest'
 
 // Helper to read test files
 const read_test_file = (filename: string): string => {
@@ -34,6 +35,65 @@ const create_test_structure = (element = `H`, atoms = 3) => ({
     properties: {},
   })),
   charge: 0,
+})
+
+describe(`Trajectory File Detection`, () => {
+  // only checking filename recognition, files don't need to exist
+  test.each([
+    // Basic trajectory files
+    [`test.traj`, true],
+    [`test.xyz`, true],
+    [`test.extxyz`, true],
+    [`test.h5`, true],
+    [`data.hdf5`, true],
+    // VASP and special files
+    [`XDATCAR`, true],
+    [`trajectory.dat`, true],
+    [`md.xyz.gz`, true],
+    [`relax.extxyz`, true],
+    // ASE ULM binary trajectory files (specific to this fix)
+    [`md_npt_300K.traj`, true],
+    [`ase-LiMnO2-chgnet-relax.traj`, true],
+    [`simulation_nvt_250K.traj`, true],
+    [`molecular_dynamics_nve.traj`, true],
+    [`water_cluster_md.traj`, true],
+    [`optimization_relax.traj`, true],
+    // Case insensitive
+    [`FILE.TRAJ`, true],
+    [`TrAjEcToRy.XyZ`, true],
+    [`trajectory`, true],
+    [`.hidden.xyz`, true],
+    // Unicode filenames
+    [`مەركەزیtrajectory.traj`, true],
+    [`日本語.xyz`, true],
+    [`file🔥emoji.h5`, true],
+    [`Мой_файл.extxyz`, true],
+    // Non-trajectory files
+    [`test.cif`, false],
+    [`test.json`, false],
+    [`random.txt`, false],
+    [`test.xyz.backup`, false],
+    // Edge cases
+    [``, false],
+    [`no.extension`, false],
+    [`.`, false],
+    [`file.xyz.`, false],
+    // Very long filename
+    [`${`a`.repeat(1000)}.xyz`, true],
+    // check the bug reported with Cr0.25Fe0.25Co0.25Ni0.25-mace-omat-qha.xyz.gz doesn't regress
+    [`Cr0.25Fe0.25Co0.25Ni0.25-mace-omat-qha.xyz`, true],
+    [`single-molecule.xyz`, true],
+    [`md-run.xyz`, true],
+    [`trajectory_data.json`, true],
+    [`md_simulation.cif`, true],
+    [`relax_output.poscar`, true],
+    [`structure.cif`, false],
+    [`molecule.json`, false],
+    [`POSCAR`, false],
+    [`data.txt`, false],
+  ])(`trajectory detection: "%s" → %s`, (filename, expected) => {
+    expect(is_trajectory_file(filename)).toBe(expected)
+  })
 })
 
 describe(`VASP XDATCAR Parser`, () => {
